requirements.txt
gym==0.25.2
matplotlib==3.5.3
numpy==1.24.2
pygame==2.1.2
PyYAML==6.0
PyYAML==6.0.1
torch==1.10.0+cu113
tqdm==4.65.0
